merchant-administration/src/main/kotlin/com/example/merchantadministration/infrastructure/query/ItemsWithSelectedAttributesCheckingQueryImpl.kt
package com.example.merchantadministration.infrastructure.query

import com.example.merchantadministration.presentation.ItemWithSelectedAttributeIdsParam
import com.example.merchantadministration.usecase.query.ItemsWithSelectedAttributesCheckingQuery

class ItemsWithSelectedAttributesCheckingQueryImpl : ItemsWithSelectedAttributesCheckingQuery {
    override fun checkItemsWithSelectedAttributes(itemWithSelectedAttributeIdsParam: ItemWithSelectedAttributeIdsParam): Boolean {
        return true
    }
}